Set timeout on read/write socket operation
EG-PMS2-LAN sometimes does not response on read requests. In this
case egctl hangs on the read call until the TCP connection times
out. The default linger value is 15, which corresponds to a
duration of approximately between 13 to 30 minutes, depending on
the retransmission timeout.

If the EG-PMS2-LAN is used in an testing environment where a DUT
is automatically powered on and off many times, the hanger
blocks the test framework unnecessary. The test framework is able
to handle failed operations. Instead blocking so long only wait for
10 seconds and report an error if we don't got any response.